[CHORE] #243: 로컬, 운영용 웹 훅 동작하지 않도록 mock 기능 구현
